Style guide

+++ contracts/PanamaFuture.sol
@@ -10,35 +10,32 @@ import "./utils/Pausable.sol";
 
 
 contract PanamaFuture is ERC20, Ownable, Pausable {
-
     using SafeMath for uint256;
 
     uint256 public currentPrice;
     ERC20 public ecoBuxAddress;
 
     // Event emitted whenever Future are transferred
-    event Transferred(
-        address owner,
-        uint256 amount
-    );
+    event Transferred(address owner, uint256 amount);
 
     // Event emitted when ECOB are transferred from user to contract
-    event EcoTransfer(
-        address owner,
-        uint256 amount
-    );
+    event EcoTransfer(address owner, uint256 amount);
 
     // Start contract with new EcoBux address
     constructor(address _ecoBuxAddress) public ERC20("PanamaFuture", "PAF") {
-        ERC20 futureAddress= ERC20(address(this));
+        ERC20 futureAddress = ERC20(address(this));
         ecoBuxAddress = ERC20(_ecoBuxAddress);
         currentPrice = 25; // Default to 1 ECOB per FUTURE. Changed by setCurrentPrice()
     }
 
     /** @dev Fuction to interface with creating and dispensing Future
-    * @param _amount Number of Future desired
-    */
-    function buyFuture(uint256 _amount) external whenNotPaused returns (uint256) {
+     * @param _amount Number of Future desired
+     */
+    function buyFuture(uint256 _amount)
+        external
+        whenNotPaused
+        returns (uint256)
+    {
         require(availableECO(msg.sender) >= _amount * currentPrice); // Require at least current price * tokens
 
         // Mint tokens and sends them to the original sender
@@ -52,24 +49,24 @@ contract PanamaFuture is ERC20, Ownable, Pausable {
     }
 
     /** @dev Function to withdraw all ETH from contract to balance
-    * @dev Users do not interact with ETH, but in case someone accidentaly sends ETH it shouldn't be stuck
-    */
+     * @dev Users do not interact with ETH, but in case someone accidentaly sends ETH it shouldn't be stuck
+     */
     function withdrawAll() external onlyOwner {
-        uint bal = address(this).balance;
+        uint256 bal = address(this).balance;
         payable(address(owner)).transfer(bal);
     }
 
     /** @dev Function to update _currentPrice
-    * @dev Throws if _currentPrice is zero
-    */
+     * @dev Throws if _currentPrice is zero
+     */
     function setCurrentPrice(uint256 _currentPrice) public onlyOwner {
         require(_currentPrice > 0); // This shouldn't ever throw, but sanitization of inputs is never a bad thing
         currentPrice = _currentPrice;
     }
 
     /** @dev Function to update ecoBuxAddress
-      * @dev Throws if _ecoBuxAddress is not a contract address
-      */
+     * @dev Throws if _ecoBuxAddress is not a contract address
+     */
     function setEcoBuxAddress(address _ecoBuxAddress) public onlyOwner {
         require(isContract(_ecoBuxAddress)); // ecoBuxAddress is common denominator contract for all subcontracts
         ecoBuxAddress = ERC20(_ecoBuxAddress);
@@ -84,14 +81,14 @@ contract PanamaFuture is ERC20, Ownable, Pausable {
     }
 
     /** @dev Function to verify user has enough EcoBux to spend
-    */
+     */
     function availableECO(address user) internal view returns (uint256) {
         return ecoBuxAddress.allowance(user, address(this));
     }
 
     /** @dev Function determine if input is contract
-      * @return bool if input is a contract
-      */
+     * @return bool if input is a contract
+     */
     function isContract(address _addr) internal view returns (bool) {
         uint32 size;
         assembly {
